requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-groq===0.1.9
python-dotenv==1.0.0
streamlit